Valid swagger specification 2.0 implementation (#3)

+++ sibapiv3/add_credits.go
@@ -13,8 +13,8 @@ package sibapiv3
 type AddCredits struct {
 
 	// SMS credits to be added to the child account
-	Sms int32 `json:"sms,omitempty"`
+	Sms int64 `json:"sms,omitempty"`
 
 	// Email credits to be added to the child account
-	Email int32 `json:"email,omitempty"`
+	Email int64 `json:"email,omitempty"`
 }
